requirements.txt
pandas
numpy
scikit-learn
dill
Pillow
tensorflow
scipy
scikeras
matplotlib
PyYAML
#-e .